Module Theory.Basic

The Basic Theory.

Implements the empty basic theory and provides a module for expressing the Basic Theory in terms of the Minimal theory.

module Make : functor (S : Minimal) -> Basic

Expresses Basic operations in terms of the Minimal operations.

module Empty : Basic

The Empty theory.