Module Float.Format

Sort describing the representation format of a floating point number.

type ('r, 's) t

val define : 'r Value.sort -> 's Bitv.t Value.sort -> ('r, 's) t Value.sort

define r s defines a sort given interpretation r of bitvector s.

val bits : ('r, 's) t Value.sort -> 's Bitv.t Value.sort

bits s returns the sort of bitvectors that are used by floating point numbers of sort s.

val exp : ('r, 's) t Value.sort -> 'r Value.sort

exp s returns an expression that describes the interpretation of the bits of the floating point numbers represented by the sort s.