Module IEEE754.Sort

IEEE754 sorts.

val define : parameters -> (('b, 'e, 't) ieee754, 's) format Float.t Value.sort

define p defines the IEEE754 format sort.

val exps : (('b, 'e, 't) ieee754, 's) format Float.t Value.sort -> 'e Bitv.t Value.sort

exps s is the sort of bitvectors for the exponent field.

val sigs : (('b, 'e, 't) ieee754, 's) format Float.t Value.sort -> 't Bitv.t Value.sort

sigs s is the sort of bitvectors for the significand field.

val bits : (('b, 'e, 't) ieee754, 's) format Float.t Value.sort -> 's Bitv.t Value.sort

bits s is the sort of bitvectors for the storage.

val spec : (('b, 'e, 't) ieee754, 's) format Float.t Value.sort -> parameters

spec s is the encoding parameters of the sort s.