Module type Grammar.Rmode

Rounding modes.

type t

an abstract type denoting a Core Theory rounding mode term.

type exp

the type for representing rounding modes in the target language.

val error : t

an ill-formed rounding mode term.

val rne : t

rne is rne.

val rtz : t

rtz is rtz.

val rtp : t

rtp is rtp.

val rtn : t

rtn is trp.

val rna : t

rna is rna.