Module Parser.Make

Make(Theory) parses AST to the specified Theory terms.

Parameters

module S : Core

Signature

val run : ('e, 'r, 's) parser -> 's list -> unit eff

run parser program the starting rule of the parser.

Applies the parser to a sequence of statements and computes a denotation of program in a Core Theory terms.