Module Unit.Compiler

Information about the compiler.

A compiler is a translator that was used to translate the code in this unit from the source to the target representation.

val name : (cls, string option) KB.slot

name the compiler name

val version : (cls, string option) KB.slot

version the compiler version.