Module Unit.Target

Information about the target architecture.

Assuming that the code is produced from source the target denotes the target for which this source is built and tailored. This module provides information about the target triplet and target features.

val arch : (cls, string option) KB.slot

arch the target architecture, e.g., arm.

val subarch : (cls, string option) KB.slot

subarch the target subarchitecture designator, e.g, v7

val vendor : (cls, string option) KB.slot

vendor the target vendor, e.g., apple

val system : (cls, string option) KB.slot

system the target operating system, e.g., darwin

val bits : (cls, int option) KB.slot

bits the number of bits in the machine word, e.g., 32

val abi : (cls, string option) KB.slot

abi target's ABI, e.g., gnueabi.

val fabi : (cls, string option) KB.slot

fabi targets floating-point ABI, e.g., hf

val cpu : (cls, string option) KB.slot

cpu the target CPU, e.g., cortex

val fpu : (cls, string option) KB.slot

fpu the target FPU.

val is_little_endian : (cls, bool option) KB.slot

is_little_endian is true if the target's default endianness is the little endian order.