Module Var.Ident

Variable identifiers.

Identifiers are compared caseless, otherwise the order loosely matches the lexicographical order of the textual representation. Identifiers of virtual variables are ordered before identifiers of physical variables and mutable virtual variables are ordered before immutable. Identifiers of a versioned variable are ordered in the ascending order of their versions. And identifiers of virtual variables are ordered in the ascending order of their numeric values, e.g., `#2` is ordered before `#123`.

before 2.4.0

the ordering was unspecified but wasn't caseless.

since

2.4.0 the ordering is caseless

type t = ident

include Core_kernel.Bin_prot.Binable.S with type Ident.t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val compare : t -> t -> int

include Ppx_sexp_conv_lib.Sexpable.S with type Ident.t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

include Core_kernel.Stringable.S with type Ident.t := t

val of_string : string -> t

val to_string : t -> string

include Base.Comparable.S with type Ident.t := t and type Ident.comparator_witness = ord

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

val clamp_exn : t -> min:t -> max:t -> t

val clamp : t -> min:t -> max:t -> t Base__.Or_error.t

type comparator_witness = ord

val comparator : (t, comparator_witness) Base__.Comparator.comparator

val validate_lbound : min:t Base__.Maybe_bound.t -> t Base__.Validate.check

val validate_ubound : max:t Base__.Maybe_bound.t -> t Base__.Validate.check

val validate_bound : min:t Base__.Maybe_bound.t -> max:t Base__.Maybe_bound.t -> t Base__.Validate.check