Module Theory.Pass

Core Theory Transformations.

A transformation pass is a simple functor. This module provides a collection of such functors and will be later extended with the transformation pipeline framework.

module Desugar (CT : Core) : Core

The Variable Desugarer.