Module type Theory.Trans

The Theory of Transcendental Functions.

val exp : rmode -> 'f float -> 'f float

exp m x is the floating-point number closest to e^x,

where b^a is b raised to the power of a and e is the base of natural logarithm.

val expm1 : rmode -> 'f float -> 'f float

expm1 m x is the floating-point number closest to e^x - 1,

where b^a is b raised to the power of a and e is the base of natural logarithm.

val exp2 : rmode -> 'f float -> 'f float

exp2 m x is the floating-point number closest to 2^x,

where b^a is b raised to the power of a.

val exp2m1 : rmode -> 'f float -> 'f float

exp2 m x is the floating-point number closest to 2^x - 1,

where b^a is b raised to the power of a.

val exp10 : rmode -> 'f float -> 'f float

exp10 m x is the floating-point number closest to 10^x,

where b^a is b raised to the power of a.

val exp10m1 : rmode -> 'f float -> 'f float

exp10m1 m x is the floating-point number closest to 10^x - 1,

where b^a is b raised to the power of a.

val log : rmode -> 'f float -> 'f float

log m x is the floating-point number closest to log x.

val log2 : rmode -> 'f float -> 'f float

log2 m x is the floating-point number closest to log x / log 2.

val log10 : rmode -> 'f float -> 'f float

log10 m x is the floating-point number closest to log x / log 10.

val logp1 : rmode -> 'f float -> 'f float

logp1 m x is the floating-point number closest to log (1+x).

val log2p1 : rmode -> 'f float -> 'f float

logp1 m x is the floating-point number closest to log (1+x) / log 2.

val log10p1 : rmode -> 'f float -> 'f float

logp1 m x is the floating-point number closest to log (1+x) / log 10.

val sin : rmode -> 'f float -> 'f float

sin m x is the floating-point number closest to sin x.

val cos : rmode -> 'f float -> 'f float

cos m x is the floating-point number closest to cos x.

val tan : rmode -> 'f float -> 'f float

tan m x is the floating-point number closest to tan x.

val sinpi : rmode -> 'f float -> 'f float

sinpi m x is the floating-point number closest to sin (pi*x).

val cospi : rmode -> 'f float -> 'f float

cospi m x is the floating-point number closest to cos (pi*x).

val atanpi : rmode -> 'f float -> 'f float

atanpi m y x is the floating-point number closest to atan(y/x) / pi.

val atan2pi : rmode -> 'f float -> 'f float -> 'f float

atanpi m y x is the floating-point number closest to atan(y/x) / (2*pi).

val asin : rmode -> 'f float -> 'f float

asin m x is the floating-point number closest to asin x.

val acos : rmode -> 'f float -> 'f float

acos m x is the floating-point number closest to acos x.

val atan : rmode -> 'f float -> 'f float

atan m x is the floating-point number closest to atan x.

val atan2 : rmode -> 'f float -> 'f float -> 'f float

atan2 m y x is the floating-point number closest to atan (y/x).

val sinh : rmode -> 'f float -> 'f float

sinh m x is the floating-point number closest to sinh x.

val cosh : rmode -> 'f float -> 'f float

cosh m x is the floating-point number closest to cosh x.

val tanh : rmode -> 'f float -> 'f float

tanh m x is the floating-point number closest to tanh x.

val asinh : rmode -> 'f float -> 'f float

asinh m x is the floating-point number closest to asinh x.

val acosh : rmode -> 'f float -> 'f float

acosh m x is the floating-point number closest to acosh x.

val atanh : rmode -> 'f float -> 'f float

atanh m x is the floating-point number closest to atanh x.